requirements.txt
Flask==3.0.3
requests==2.31.0
chromadb==0.4.24
PyPDF2==3.0.1
python-pptx==0.6.23
werkzeug==3.0.1
